LLM NPC base (LLMNpcMixin/LLMNpc) + companion archetype
Standing up a non-bartender LLM NPC surfaced that the engagement + agentic
tool loops lived on Bartender. Extract them into a reusable brain.

- New typeclasses/llm_npc.py: LLMNpcMixin (at_msg_receive, classify,
  _try_llm_reply, the agentic loop, render, short-term memory, perception)
  + LLMNpc(LLMNpcMixin, Character) — a generic LLM-driven social NPC whose
  job is its persona archetype. Subclass hooks: _handle_directed_speech,
  _run_context_tool, _handle_action_tool, _name_aliases.
- Bartender = LLMNpcMixin + Character + bartending only (orders/gratitude
  via _handle_directed_speech; check_stock/prepare_drink via the tool
  hooks). No behaviour change — the bar tests pin Sable's routing.
- Commit the approved `companion` archetype to ARCHETYPES (social,
  tools=[], BASE look only).
- test_bar patches retargeted to the llm_npc namespace (patch-where-used).

97 bar + prompt tests green.

Closes #736
