+++ typeclasses/bar.py
@@ -13,9 +13,7 @@ v1 is intentionally lenient where the spec defers (ownership gating, recipe-save
 UX) so the loop is testable; those are later slices.
 """
 
-import json
 import random
-from functools import partial
 from time import monotonic
 
 from evennia import CmdSet
@@ -24,12 +22,8 @@ from evennia.utils import delay
 
 from typeclasses.items import Item
 from typeclasses.characters import Character
-from typeclasses.llm_persona import build_persona
+from typeclasses.llm_npc import LLMNpcMixin
 from world.grammar import capitalize_first, with_article
-from world.llm.client import llm_enabled, request_turn
-from world.llm.prompt import (
-    CONTEXT_TOOLS, build_messages, parse_turn, schema_for, tool_names,
-)
 from world.shop.utils import format_currency
 from world.bar import (
     DEFAULT_BAR_SNACKS,
@@ -334,20 +328,14 @@ ACK_EMOTES = (
 #: Minimum seconds between acknowledgements, so a chatty room doesn't spam them.
 ACK_COOLDOWN = 6.0
 
-#: LLM-driven conversational replies (LLM_GAMEMASTER_SPEC Phase 1, #707). Gated
-#: by BOTH ``settings.LLM_GM_ENABLED`` (deployment) and ``npc.db.llm_driven``
-#: (per-NPC), so a scripted bartender stays byte-identical when the LLM is off.
-LLM_DIRECTED_COOLDOWN = 4.0    # min seconds between replies to direct address/name
-LLM_AMBIENT_COOLDOWN = 45.0    # she rarely volunteers into overheard chatter
-LLM_AMBIENT_CHANCE = 0.35      # ...and not every eligible time
-LLM_HISTORY_TURNS = 6          # recent turns kept per interlocutor (anti-repetition)
-#: CONTEXT_TOOLS (read-only → loop the result back; vs action tools → real command)
-#: is derived from the prompt-layer tool registry, so they can't desync.
-LLM_MAX_TOOL_ROUNDS = 3        # cap the agentic loop so it can't spin forever
 
+class Bartender(LLMNpcMixin, Character):
+    """An NPC that takes drink orders by being spoken to (the `to` command).
 
-class Bartender(Character):
-    """An NPC that takes drink orders by being spoken to (the `to` command)."""
+    Bartending mechanics on top of the shared LLM brain (``LLMNpcMixin``):
+    orders/gratitude intercept speech before the LLM layer, and the
+    ``check_stock`` / ``prepare_drink`` tools route to the bar.
+    """
 
     def at_object_creation(self):
         super().at_object_creation()
@@ -376,44 +364,22 @@ class Bartender(Character):
                 return obj
         return None
 
-    def at_msg_receive(self, text=None, from_obj=None, **kwargs):
-        """React to speech nearby, via the shared speech backbone.
-
-        Every verb that carries words — ``say``, ``to``, or a pose with an
-        embedded quote — delivers the same structured payload (``speech`` =
-        the words, ``addressed`` = whether this bartender was the one spoken to),
-        and only to listeners who can hear. So a single check handles all three:
-
-          - **Gratitude** (thanks/cheers/much obliged...) from any heard speech —
-            answered with a small non-verbal acknowledgement. Checked first so
-            "thanks for the rotgut" reads as a thank-you, not a re-order.
-          - **A directed order** — speech that was *addressed* to this bartender
-            (``to <bartender> ...`` or ``.slide up to <bartender>, "..."``) is
-            matched against the menu and made.
+    def _handle_directed_speech(self, speech, speaker, kwargs):
+        """Bartender intercept (before the LLM layer): gratitude → a small
+        non-verbal nod (checked first so "thanks for the rotgut" reads as a
+        thank-you, not a re-order); an *addressed* line → a menu order. Returns
+        True when handled, so conversation only reaches the LLM otherwise.
         """
-        speech = kwargs.get("speech")
-        speaker = from_obj
-        if not speech or speaker is None or speaker is self:
-            return True
-
         if self._is_gratitude(speech):
             self._acknowledge()
             return True
-
         if kwargs.get("addressed"):
             delay(1.5, self._fulfil_order, speech, speaker)
             return True
+        return False
 
-        # LLM conversational layer — only when both gates are on, so the
-        # scripted NPC is byte-identical with the LLM off. Orders and gratitude
-        # above are never routed here.
-        if self.db.llm_driven and llm_enabled():
-            kind = self._classify_speech(speech, speaker)
-            if kind == "directed":
-                delay(1.5, self._try_llm_reply, speech, speaker, "directed")
-            elif kind == "ambient":
-                delay(1.0, self._try_llm_reply, speech, speaker, "ambient")
-        return True
+    def _name_aliases(self):
+        return ["bartender", "barkeep", "barkeeper"]
 
     @staticmethod
     def _is_gratitude(content):
@@ -470,208 +436,25 @@ class Bartender(Character):
             f"and {closer}."
         )
 
-    # --- LLM-driven conversation (gated; LLM_GAMEMASTER_SPEC Phase 1) --------
-
-    def _classify_speech(self, speech, speaker):
-        """Cheap reactor-side gate: ``directed`` | ``ambient`` | ``ignore``."""
-        # Loop guard: never react to another NPC's broadcast speech, so two
-        # LLM-driven NPCs can't ping-pong on ambient lines.
-        if (getattr(speaker.db, "is_bartender_npc", False)
-                or getattr(speaker.db, "llm_driven", False)):
-            return "ignore"
-        # Named, or effectively alone with the speaker (then any line is plainly
-        # for her), counts as directed; otherwise it's overheard chatter.
-        if self._mentions_self(speech) or self._is_alone_with(speaker):
-            return "directed"
-        return "ambient"
-
-    def _is_alone_with(self, speaker):
-        """True when no other character shares the room — so the speaker can only
-        be talking to this NPC."""
-        if not self.location:
-            return False
-        return not any(
-            o is not self and o is not speaker and isinstance(o, Character)
-            for o in self.location.contents
-        )
-
-    def _mentions_self(self, speech):
-        """Whether a line names this bartender (key, keyword, or generic role)."""
-        low = (speech or "").lower()
-        names = [self.key.lower()]
-        if self.sdesc_keyword:
-            names.append(self.sdesc_keyword.lower())
-        names += ["bartender", "barkeep", "barkeeper"]
-        return any(n and n in low for n in names)
-
-    def _try_llm_reply(self, line, patron, mode, on_fail=None):
-        """Route a conversational line to the LLM sidecar, off the reactor.
-
-        Returns True if the LLM path was taken (caller suppresses any scripted
-        fallback), False if gated/throttled off so the caller can fall back.
-        """
-        if not self.db.llm_driven or not llm_enabled():
-            return False
-        if (not self.location
-                or getattr(patron, "location", None) is not self.location):
-            return False
-
-        now = monotonic()
-        last = self.ndb.last_llm or 0
-        if mode == "ambient":
-            if now - last < LLM_AMBIENT_COOLDOWN:
-                return True  # throttled: stay silent rather than spam
-            if random.random() > LLM_AMBIENT_CHANCE:
-                return True  # eligible, but she didn't bite this time
-        elif now - last < LLM_DIRECTED_COOLDOWN:
-            return True
-        self.ndb.last_llm = now
-
-        # Capture everything reactor-side BEFORE threading (SQLite/Evennia-thread
-        # contract). The agentic loop re-calls from the reactor-side callback.
-        persona = build_persona(self)
-        speaker_name = patron.get_display_name(self)
-        perception = self._perceive(patron)
-        history = self._recent_history(patron)
-        messages = build_messages(persona, speaker_name, line or "", mode,
-                                  perception, history)
-        self._agentic_round(messages, persona, patron, line or "", speaker_name,
-                            on_fail or self._llm_silent, rounds=0)
-        return True
-
-    def _perceive(self, patron):
-        """What this NPC sees when it looks at the patron — grounds the model's
-        description so it can't invent the speaker's appearance. ANSI-stripped,
-        identity-gated, trimmed to a sentence-bounded summary (the full
-        return_appearance bloats context and truncates mid-word)."""
-        try:
-            from evennia.utils.ansi import strip_ansi
-            raw = patron.return_appearance(self)
-            if not raw:
-                return None
-            text = " ".join(strip_ansi(raw).split())
-            if len(text) > 300:
-                cut = text[:300]
-                for end in (". ", "! ", "? "):
-                    i = cut.rfind(end)
-                    if i > 120:
-                        cut = cut[: i + 1]
-                        break
-                text = cut.rstrip()
-            return text
-        except Exception:
-            return None
-
-    # --- short-term conversation memory (per interlocutor, ndb/ephemeral) ----
-
-    @staticmethod
-    def _hist_key(patron):
-        return f"#{patron.id}"
-
-    def _recent_history(self, patron):
-        """The recent turns with this interlocutor — fed back into the prompt so
-        the model sees what it just said and stops repeating itself."""
-        return (self.ndb.llm_history or {}).get(self._hist_key(patron), [])
-
-    # --- the agentic tool loop (constrained turn → context tools → reply) ----
-
-    def _agentic_round(self, messages, persona, patron, line, speaker_name,
-                       on_fail, rounds):
-        """One constrained generation; the reactor-side callback either runs a
-        context tool and loops, or renders the final reply + any action tool."""
-        request_turn(
-            messages,
-            on_turn=partial(self._on_turn, messages, persona, patron, line,
-                            speaker_name, on_fail, rounds),
-            on_fail=on_fail,
-            schema=schema_for(persona),  # tool enum scoped to the archetype
-        )
-
-    def _on_turn(self, messages, persona, patron, line, speaker_name, on_fail,
-                 rounds, raw):
-        turn = parse_turn(raw, persona, tool_names(persona))
-        tool, arg = turn["tool"], turn["tool_argument"]
-        # Context tool: run the real read, feed the result back, loop.
-        if tool in CONTEXT_TOOLS and rounds < LLM_MAX_TOOL_ROUNDS:
-            result = self._run_context_tool(tool, arg, patron)
-            extended = messages + [
-                {"role": "assistant",
-                 "content": raw if isinstance(raw, str) else json.dumps(raw)},
-                {"role": "user", "content": f"[tool result · {tool}] {result}"},
-            ]
-            self._agentic_round(extended, persona, patron, line, speaker_name,
-                                on_fail, rounds + 1)
-            return
-        # Terminal: render speech/action, run the action tool, remember.
-        self._render_llm_reply(turn["speech"], turn["action"])
-        if tool == "prepare_drink" and arg and self.location:
-            self.execute_cmd(f"prepare {arg}")
-        self._remember_turn(patron, line, speaker_name, turn["speech"],
-                            turn["action"])
+    # --- bartender tool routing (over the shared LLM brain) ------------------
 
     def _run_context_tool(self, tool, arg, patron):
-        """Run a read-only context tool via the game's real logic and return its
-        result for the model. (Reads only — world-mutating tools go through real
-        commands in ``_on_turn``.)"""
-        if tool == "look":
-            return self._perceive(patron) or "nothing remarkable"
+        """Extend the read-only tools with ``check_stock`` (``look`` is the
+        mixin's). Reads only — the action tool goes through a real command."""
         if tool == "check_stock":
             bar = self._find_bar()
             menu = (bar.db.menu if bar else None) or self.db.menu or []
             names = [r.get("name") for r in menu if r.get("name")]
             return ("Serves: " + ", ".join(names)) if names else "nothing on tap"
-        return ""
+        return super()._run_context_tool(tool, arg, patron)
 
-    def _remember_turn(self, patron, line, speaker_name, speech, action):
-        """Append the rendered turn to short-term memory (anti-repetition)."""
-        reply = self._reconstruct_reply(speech, action)
-        if not reply:
-            return
-        hist = self.ndb.llm_history or {}
-        key = self._hist_key(patron)
-        turns = list(hist.get(key, []))
-        turns.append({
-            "user": f'{speaker_name} says to you: "{line}"',
-            "assistant": reply,
-        })
-        hist[key] = turns[-LLM_HISTORY_TURNS:]
-        self.ndb.llm_history = hist
-
-    @staticmethod
-    def _reconstruct_reply(speech, action):
-        """Re-form the model's own reply for the history (so it sees its prior
-        gestures and phrasing, in the same format it produces)."""
-        if action and speech:
-            return f'*{action}* "{speech}"'
-        if action:
-            return f"*{action}*"
-        if speech:
-            return f'"{speech}"'
-        return ""
-
-    def _render_llm_reply(self, speech, action):
-        """Render the sidecar reply as ONE fluid emote — the MUD-native way to act
-        and speak in a single beat. The embedded quote rides the hearing-gated
-        speech rails (``tokenize_emote`` → ``SpeechToken``) and character refs in
-        the action resolve per-observer. Falls back to a bare pose or say."""
-        if not self.location:
-            return
-        speech = speech.strip().strip('"').strip() if speech else None
-        action = action.strip() if action else None
-        if action and speech:
-            if action[-1] not in ".!?…,":
-                action += "."
-            self.execute_cmd(f'pose {action} "{speech}"')
-        elif action:
-            self.execute_cmd(f"pose {action}")
-        elif speech:
-            self.execute_cmd(f"say {speech}")
+    def _handle_action_tool(self, tool, arg, patron):
+        """Route the ``prepare_drink`` action tool to the real ``prepare``
+        command — the bar makes the drink for real (never a faked pour)."""
+        if tool == "prepare_drink" and arg and self.location:
+            self.execute_cmd(f"prepare {arg}")
 
     def _llm_fallback(self):
         """Sidecar failed on an addressed non-order: the curt scripted line."""
         if self.location:
             self.execute_cmd("say Don't serve that here.")
-
-    def _llm_silent(self):
-        """Sidecar failed or declined on conversation: stay quiet."""
-        return None
